+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - Update to rails 6.0.1
   - Update to carrierwave 2.0.2
   - Update to devise 4.7.1
+  - Update to font-awesome-sass 5.11.2
   - Update to haml 5.1.2
   - Update to jquery-rails 4.3.5